requirements.txt
# FastAPI NAS backend — Pi-friendly pins; bump as needed
fastapi>=0.109.0,<0.116.0
uvicorn[standard]>=0.27.0,<0.33.0
motor>=3.3.0,<3.8.0
certifi>=2024.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
bcrypt>=4.1.0,<5.0.0
python-multipart>=0.0.6,<0.1.0
email-validator>=2.1.0,<3.0.0
httpx>=0.26.0,<0.29.0
